requirements.txt
# Google Cloud & ADK
google-cloud-aiplatform>=1.70.0
google-adk>=1.23.0
vertexai>=1.43.0
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.18.0
google-genai>=0.3.0
google-auth>=2.35.0

# Gemini API
google-generativeai>=0.8.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Web framework
streamlit>=1.39.0
plotly>=5.24.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.4.0
requests>=2.32.0

# Testing
pytest>=8.3.0
pytest-cov>=6.0.0
pytest-asyncio>=0.24.0

# Development
black>=24.8.0
flake8>=7.1.0
mypy>=1.11.0
